refactor(ImageDetection): extract BoxAnnotations component
